Tidy up data preprocessing https://github.com/transferwise/auto-causality/issues/95
