+++ auto_causality/datasets.py
@@ -1,6 +1,4 @@
 import pandas as pd
-import numpy as np
-from auto_causality.utils import featurize
 
 
 def nhefs() -> pd.DataFrame:
@@ -237,41 +235,3 @@ def synth_acic(condition=1) -> pd.DataFrame:
     data.rename(columns={"z": "treatment"}, inplace=True)
 
     return data
-
-
-def preprocess_dataset(
-    data: pd.DataFrame, treatment="treatment", targets=["y_factual"]
-) -> tuple:
-    """preprocesses dataset for causal inference
-
-    Args:
-        data (pd.DataFrame): a dataset for causal inference
-
-    Returns:
-        tuple: dataset, features_x, features_w, list of targets, name of treatment
-    """
-    if (treatment not in data.columns) or any(t not in data.columns for t in targets):
-        print("some of the requested variables are not part of this dataset.")
-        print(f"available columns are {[c for c in data.colums if 'x' not in c]}.")
-        return None
-    else:
-        # prepare the data
-        features = [c for c in data.columns if c not in [treatment] + targets]
-
-        data[treatment] = data[treatment].astype(int)
-        # this is a trick to bypass some DoWhy/EconML bugs
-        data["random"] = np.random.randint(0, 2, size=len(data))
-
-        used_df = featurize(
-            data,
-            features=features,
-            exclude_cols=[treatment] + targets,
-            drop_first=False,
-        )
-        used_features = [c for c in used_df.columns if c not in [treatment] + targets]
-
-        # Let's treat all features as effect modifiers
-        features_X = [f for f in used_features if f != "random"]
-        features_W = [f for f in used_features if f not in features_X]
-
-        return used_df, features_X, features_W, targets, treatment
